Rename float to double

+++ tests/semantics/03-assign-incompatible-types-02.pas
@@ -1,6 +1,6 @@
 var
     a: Integer;
-    f: float; 
+    f: double; 
 begin
     a /= f;
 end.
