+++ tests/semantics/11-simple-unary-not-01.pas
@@ -1,5 +1,5 @@
 var
-    b : float;
+    b : double;
     
 begin
     b:= not 2.2;
