+++ tests/semantics/19-variable-initialize-01.pas
@@ -1,6 +1,6 @@
 var
     a : Integer = 2 * 8 + 4; // 20
-    b : Float = 2. + 12 / 5; // 4.4
-    c : Float = 20 - 100;
+    b : double = 2. + 12 / 5; // 4.4
+    c : double = 20 - 100;
 begin
 end.
