+++ tests/semantics/20-array-access-05.pas
@@ -1,6 +1,6 @@
 var 
-    a : Array[1..100] of float;
-    b : float;
+    a : Array[1..100] of double;
+    b : double;
 begin
     
     b := a[20*2 - 7];
