+++ tests/semantics/24-record-delcaration-05.pas
@@ -3,7 +3,7 @@ type
     end;
     
     sample = record
-        float : integer = 3;
+        double : integer = 3;
         char : integer;
     end;
